pyproject.toml
[tool.poetry]
name = "coffeegis-telegram-bot"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
python-telegram-bot = "^21.0.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
grpcio = "^1.62.1"
grpcio-tools = "^1.62.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.1"
prometheus-fastapi-instrumentator = "^6.1.0"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pylint = "^3.1.0"
flake8 = "^7.0.0"
ruff = "^0.3.1"


[tool.poetry.group.migration.dependencies]
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
